refactor: Handler 추상화
