+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,28 +1,24 @@
 package nextstep.mvc;
 
 import nextstep.exception.NoSuchHandlerException;
-import nextstep.mvc.asis.Controller;
-import nextstep.mvc.tobe.HandlerExecution;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.Handler;
+import nextstep.mvc.tobe.ViewResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
 @WebServlet(name = "dispatcher", urlPatterns = "/", loadOnStartup = 1)
 public class DispatcherServlet extends HttpServlet {
-    private static final long serialVersionUID = 1L;
     private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
-    private static final String DEFAULT_REDIRECT_PREFIX = "redirect:";
+    private static final long serialVersionUID = 1L;
 
     private List<HandlerMapping> mappings;
 
@@ -36,43 +32,24 @@ public class DispatcherServlet extends HttpServlet {
     }
 
     @Override
-    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException {
         logger.debug("Method : {}, Request URI : {}", req.getMethod(), req.getRequestURI());
-
-        Object handler = getHandler(req);
+        Handler handler = getHandler(req);
 
         try {
-            if (handler instanceof Controller) {
-                String viewName = ((Controller) handler).execute(req, resp);
-                move(viewName, req, resp);
-                return;
-            }
-            if (handler instanceof HandlerExecution) {
-                ModelAndView mav = ((HandlerExecution) handler).handle(req, resp);
-                mav.getView().render(mav.getModel(), req, resp);
-            }
+            Object view = handler.execute(req, resp);
+            new ViewResolver().resolve(view, req, resp);
         } catch (Throwable e) {
             logger.error("Exception : {}", e.getMessage());
             throw new ServletException(e.getMessage());
         }
     }
 
-    private Object getHandler(HttpServletRequest req) {
+    private Handler getHandler(HttpServletRequest req) {
         return mappings.stream()
                 .map(m -> m.getHandler(req))
                 .filter(Objects::nonNull)
                 .findFirst()
                 .orElseThrow(NoSuchHandlerException::new);
-    }
-
-    private void move(String viewName, HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        if (viewName.startsWith(DEFAULT_REDIRECT_PREFIX)) {
-            resp.sendRedirect(viewName.substring(DEFAULT_REDIRECT_PREFIX.length()));
-            return;
-        }
-
-        RequestDispatcher rd = req.getRequestDispatcher(viewName);
-        rd.forward(req, resp);
     }
 }
