+++ nextstep-mvc/src/main/java/nextstep/mvc/asis/Controller.java
@@ -1,8 +1,11 @@
 package nextstep.mvc.asis;
 
+import nextstep.mvc.tobe.Handler;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public interface Controller {
+public interface Controller extends Handler {
+    @Override
     String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception;
 }
