+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/HandlerExecution.java
@@ -4,6 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @FunctionalInterface
-public interface HandlerExecution {
-    ModelAndView handle(HttpServletRequest request, HttpServletResponse response) throws Exception;
+public interface HandlerExecution extends Handler {
+    @Override
+    ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws Exception;
 }
